add method DFSChunk to QuadTree

+++ Agario/src/main/java/com/example/agario/utils/Dimension.java
@@ -1,35 +1,35 @@
 package com.example.agario.utils;
 
 public class Dimension {
-    private int xMin;
-    private int yMin;
-    private int xMax;
-    private int yMax;
+    private double xMin;
+    private double yMin;
+    private double xMax;
+    private double yMax;
 
-    public Dimension(int xMin, int yMin, int xMax, int yMax) {
+    public Dimension(double xMin, double yMin, double xMax, double yMax) {
         this.xMin = xMin;
         this.yMin = yMin;
         this.xMax = xMax;
         this.yMax = yMax;
     }
 
-    public int getxMin() {
+    public double getxMin() {
         return xMin;
     }
 
-    public int getyMin() {
+    public double getyMin() {
         return yMin;
     }
 
-    public int getxMax() {
+    public double getxMax() {
         return xMax;
     }
 
-    public int getyMax() {
+    public double getyMax() {
         return yMax;
     }
 
-    public boolean inRange(int x, int y) {
+    public boolean inRange(double x, double y) {
         return (x >= this.getxMin() && x <= this.getxMax() && y >= this.getyMin() && y <= this.getyMax());
     }
 }
